feat(core): extract plan-id + plan-hu-ops to @karajan/core (KJC-TSK-0511)
PR5 of KJC-TSK-0511. Moves the HU CRUD helpers (addHu, removeHu,
updateHu, updateHuStatus, certifyAllHus, ...) plus the plan/HU id
generators to @karajan/core so @karajan/hu-board can stop reaching
into ../../../src.

- packages/core/src/plan-id.js, plan-hu-ops.js: verbatim copies.
- src/plan/plan-id.js, plan-hu-ops.js: now thin shims re-exporting
  from @karajan/core/* (CLI callers unchanged).
- packages/hu-board/src/plan-mutations.js: import from
  @karajan/core/plan-hu-ops instead of the relative src path.
- 10 new tests in packages/core/tests covering id generation +
  HU CRUD invariants (94/94 CLI consumer tests still green,
  49/49 hu-board plan-mutations tests still green).

Net delta +116 LOC (well under the 200 budget).
