+++ packages/core/README.md
@@ -22,6 +22,8 @@ here.
 | `@karajan/core/paths` | `resolveHome`, `getKarajanHome`, `getSessionRoot`, `getSonarComposePath`, `getOllamaComposePath`, `getWebperfDir`, `getRunsDir`, `getPromptsDir`, `__resetKjHomeWarningForTests` |
 | `@karajan/core/run-registry` | `runsDir`, `registerRun`, `unregisterRun`, `listActiveRuns` |
 | `@karajan/core/vec-store` | `dbPath`, `openVecStore`, `insertChunk`, `searchSimilar`, `searchBM25`, `projectSlug`, `deleteChunksBySource`, `findChunkByHash`, `getEmbeddingsByIds`, `getLastIndexedCommit`, `setLastIndexedCommit`, `countChunks` |
+| `@karajan/core/plan-id` | `generatePlanId`, `generateHuId`, `normaliseAlias` |
+| `@karajan/core/plan-hu-ops` | `addHu`, `removeHu`, `updateHu`, `updateHuStatus`, `setHuOutcome`, `setPlanOutcome`, `autoCertifyPendingHus`, `assertPlanRunnable`, `computePlanOutcome`, `certifyAllHus`, `reorderHus` |
 
 The root export (`@karajan/core`) re-exports everything via the barrel
 in `src/index.js`, but prefer the subpath form so the bundler can
@@ -29,5 +31,6 @@ tree-shake unused modules.
 
 ## Status
 
-KJC-TSK-0511 PR4 — `vec-store` extracted (with `better-sqlite3` +
-`sqlite-vec` deps). See the task card for the full roadmap.
+KJC-TSK-0511 PR5 — `plan-id` + `plan-hu-ops` extracted (HU CRUD on v2
+plans, plus globally-unique plan/HU id generation). See the task card
+for the full roadmap.
